package renamed from osdconfigapi to api
